+++ tests/set_up_database.py
@@ -4,10 +4,13 @@ data. I usually keep one mouse: `test_mouse`, id 6094f7212597df357fa24a8c
 
 This file just does that.
 """
-from pathlib import Path
+
 import json
-import flexilims as flm
+from pathlib import Path
+
 from flexiznam.config.config_tools import get_password
+
+import flexilims as flm
 
 USERNAME = "blota"
 password = get_password(username=USERNAME, app="flexilims")
@@ -24,7 +27,9 @@ def delete_recursive(flm_sess, current_id):
     print(f"Deleting {len(children)} children of {current_id}")
     for c in children:
         delete_recursive(flm_sess, c["id"])
-        flm_sess.delete(c["id"])
+        rep = flm_sess.delete(c["id"])
+        if rep != "deleted successfully [1, 0]":
+            print(f"Error deleting {c['id']}: {rep}")
 
 
 print("Deleting all entities below test mouse")